requirements.txt
requests==2.32.3
python-dotenv==1.0.1
pyyaml==6.0.2
pytest==8.3.5
streamlit>=1.30.0
pandas>=2.0.0
